requirements.txt
python-dotenv
discord.py
uuid6
aiofiles